feat(netplay): animation sprites, visual effects, FPS plugin updates
- Extract Mosquiton pure simulation into fps_core (MosquitonSimState,
  MosquitonSimConfig, tick_mosquiton_sim) — single source of truth for
  combat state machine, cooldowns, and attack decisions
- Server enemy attacks now use fps_core sim path (ServerMosquitonSim
  replaces EnemyAttackCooldown), gaining strafing + recovery states
- Unify all Mosquiton config constants: SP MosquitonConfig::default()
  derives from fps_core::config (shoot/melee cooldown, ranges, speeds)
- fps_core → fps convergence: re-export camera, collision, map, raycast,
  enemy types; fps Mosquiton delegates to fps_core sim
- Remote flamethrower: flame billboards with wall clipping + impact,
  FlameCharMark wall char decals, RemoteFlameStates cleanup
- TurnChordState: hold-to-select, release-to-confirm with 150ms window,
  pre-held direction blocking, 18 chord tests
- SP shoot animation: Mosquiton projectile spawn delayed to cue point
- Damage flicker frequency doubled
- Flamethrower collision: line-distance model replaces cone angle
- Wall impact sprite: uses dedicated flame_wall_hit animation
- Flame lifecycle cleanup: FlameCharCooldowns, FlameActive(false) on
  death/disconnect, BurnContactCooldowns cleared on death
- Hardened test port allocation with bind-check retry
- Authority regression tests: dead player fire, out-of-order sequence,
  cooldown state verification
- TickPlugin Time<Fixed> guard for MinimalPlugins safety
- register_net_all debug_assert for RepliconChannels ordering
- Regenerated weapon sprites (gun, flamethrower, muzzle flash)
- Muzzle flash: offset adjusted, animation speed increased (250ms)

+++ crates/carcinisation_cutscene/src/data.rs
@@ -394,7 +394,7 @@ pub struct LetterboxMoveEvent {
 impl LetterboxMoveEvent {
     #[must_use]
     pub fn open() -> Self {
-        Self::new(LETTERBOX_NORMAL_SPEED, LETTERBOX_HEIGHT as f32)
+        Self::new(LETTERBOX_NORMAL_SPEED, LETTERBOX_HEIGHT)
     }
 
     #[must_use]
@@ -404,7 +404,7 @@ impl LetterboxMoveEvent {
 
     #[must_use]
     pub fn show() -> Self {
-        Self::new(LETTERBOX_INSTANT_SPEED, LETTERBOX_HEIGHT as f32)
+        Self::new(LETTERBOX_INSTANT_SPEED, LETTERBOX_HEIGHT)
     }
 
     #[must_use]
@@ -439,4 +439,4 @@ impl From<LetterboxMove> for LetterboxMoveEvent {
 // Constants for letterbox. Defined here so both app and editor can use them.
 pub const LETTERBOX_NORMAL_SPEED: f32 = 10.;
 pub const LETTERBOX_INSTANT_SPEED: f32 = f32::MAX;
-pub const LETTERBOX_HEIGHT: u32 = 20;
+pub const LETTERBOX_HEIGHT: f32 = 20.0;
